requirements.txt
# Terminal AI Chatbot - Dependencies (OpenAI)
# Python 3.10+

langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
